+++ Tests/Set-CIisAppPoolProcessModel.Tests.ps1
@@ -1,4 +1,5 @@
-using module '..\Carbon.Iis'
+
+using module '..\Carbon.IIS\Carbon.IIS.Enums.psm1'
 using namespace Microsoft.Web.Administration
 
 Set-StrictMode -Version 'Latest'
@@ -8,30 +9,33 @@ BeforeAll {
 
     $script:testNum = 0
 
+    #
     $script:defaultDefaults = @{}
-    (Get-CIisWebsite -Defaults).LogFile.Attributes |
+    (Set-CIisAppPoolProcessModel -AsDefaults).ProcessModel.Attributes |
         Where-Object 'IsInheritedFromDefaultValue' -EQ $false |
         ForEach-Object { $script:defaultDefaults[$_.Name] = $_.Value }
 
     # All non-default values.
     $script:nonDefaultArgs = @{
-        CustomLogPluginClsid = '931a0831-2301-4a0b-9887-ee9a7d0c10df';
-        Directory = 'C:\my\log\files';
-        Enabled = $false;
-        FlushByEntryCountW3CLog = 1000;
-        LocalTimeRollover = $true;
-        LogExtFileFlags = ([LogExtFileFlags]::Date -bor [LogExtFileFlags]::Host);
-        LogFormat = [LogFormat]::Custom;
-        LogSiteID = $true;
-        LogTargetW3C = ([LogTargetW3C]::File -bor [LogTargetW3C]::ETW);
-        MaxLogLineLength = 2000;
-        Period = [LoggingRolloverPeriod]::Hourly;
-        TruncateSize = 1048576;
+        'identityType' = [ProcessModelIdentityType]::ApplicationPoolIdentity;
+        'idleTimeout' = [TimeSpan]'00:00:00';
+        'idleTimeoutAction' = [IdleTimeoutAction]::Suspend;
+        'loadUserProfile' = $true;
+        'logEventOnProcessModel' = [ProcessModelLogEventOnProcessModel]::None;
+        'logonType' = [CIisProcessModelLogonType]::Service;
+        'manualGroupMembership' = $true;
+        'maxProcesses' = [UInt32]2;
+        'pingingEnabled' = $true;
+        'pingInterval' = [TimeSpan]'00:00:10';
+        'pingResponseTime' = [TimeSpan]'00:05:00';
+        'requestQueueDelegatorIdentity' = 'SYSTEM';
+        'setProfileEnvironment' = $false;
+        'shutdownTimeLimit' = [TimeSpan]'00:00:30';
+        'startupTimeLimit' = [TimeSpan]'00:05:00';
     }
 
     # Sometimes the default values in the schema aren't quite the default values.
     $script:notQuiteDefaultValues = @{
-        Directory = '%SystemDrive%\inetpub\logs\LogFiles';
     }
 
     function ThenDefaultsSetTo
@@ -50,13 +54,14 @@ BeforeAll {
         param(
             [Parameter(Position=0)]
             [hashtable] $Values = @{},
+
             [switch] $OnDefaults
         )
 
-        $website = Get-CIisWebsite -Name $script:websiteName
-        $website | Should -Not -BeNullOrEmpty
+        $targetParent = Get-CIisAppPool -Name $script:appPoolName -Defaults:$OnDefaults
+        $targetParent | Should -Not -BeNullOrEmpty
 
-        $target = $website.LogFile
+        $target = $targetParent.ProcessModel
         $target | Should -Not -BeNullOrEmpty
 
         $asDefaultsMsg = ''
@@ -66,7 +71,6 @@ BeforeAll {
         }
         foreach( $attr in $target.Schema.AttributeSchemas )
         {
-            $currentValue = $target.GetAttributeValue($attr.Name)
             $expectedValue = $attr.DefaultValue
             $becauseMsg = 'default'
             if( $script:notQuiteDefaultValues.ContainsKey($attr.Name))
@@ -80,7 +84,17 @@ BeforeAll {
                 $becauseMsg = 'custom'
             }
 
-            if( $currentValue -is [TimeSpan] )
+            if( $expectedValue -is [securestring] )
+            {
+                $expectedValue = [pscredential]::New('ignoded', $expectedValue).GetNetworkCredential().Password
+            }
+
+            $currentValue = $target.GetAttributeValue($attr.Name)
+            if( $currentValue -is [securestring] )
+            {
+                $currentValue = [pscredential]::New('ignorded', $currentValue).GetNetworkCredential().Password
+            }
+            elseif( $currentValue -is [TimeSpan] )
             {
                 if( $expectedValue -match '^\d+$' )
                 {
@@ -98,69 +112,67 @@ BeforeAll {
     }
 }
 
-Describe 'Set-CIisWebsiteLogFile' {
+Describe 'Set-CIisAppPoolProcessModel' {
     BeforeAll {
         Start-W3ServiceTestFixture
-        Install-CIisAppPool -Name 'Set-CIisWebsiteLogFile'
     }
 
     AfterAll {
-        Uninstall-CIisAppPool -Name 'Set-CIisWebsiteLogFile'
         Complete-W3ServiceTestFixture
     }
 
     BeforeEach {
-        $script:websiteName = "Set-CIisWebsiteLogFile$($script:testNum++)"
-        Set-CIisWebsiteLogFile -AsDefaults @script:defaultDefaults
-        Install-CIisWebsite -Name $script:websiteName -PhysicalPath (New-TestDirectory) -AppPoolName $script:websiteName
+        $script:appPoolName = "Set-CIisAppPoolProcessModel$($script:testNum++)"
+        Set-CIisAppPoolProcessModel -AsDefaults @script:defaultDefaults
+        Install-CIisAppPool -Name $script:appPoolName
     }
 
     AfterEach {
-        Uninstall-CIisWebsite -Name $script:websiteName
-        Set-CIisWebsiteLogFile -AsDefaults @script:defaultDefaults
+        Uninstall-CIisAppPool -Name $script:appPoolName
+        Set-CIisAppPoolProcessModel -AsDefaults @script:defaultDefaults
     }
 
     It 'should set and reset all values' {
         $infos = @()
-        Set-CIisWebsiteLogFile -SiteName $script:websiteName @nonDefaultArgs -InformationVariable 'infos'
+        Set-CIisAppPoolProcessModel -AppPoolName $script:appPoolName @nonDefaultArgs -InformationVariable 'infos'
         $infos | Should -Not -BeNullOrEmpty
         ThenHasValues $nonDefaultArgs
 
         # Make sure no information messages get written because no changes are being made.
-        Set-CIisWebsiteLogFile -SiteName $script:websiteName @nonDefaultArgs -InformationVariable 'infos'
+        Set-CIisAppPoolProcessModel -AppPoolName $script:appPoolName @nonDefaultArgs -InformationVariable 'infos'
         $infos | Should -BeNullOrEmpty
         ThenHasValues $nonDefaultArgs
 
-        Set-CIisWebsiteLogFile -SiteName $script:websiteName
+        Set-CIisAppPoolProcessModel -AppPoolName $script:appPoolName
         ThenHasDefaultValues
     }
 
     It 'should support WhatIf when updating all values' {
-        Set-CIisWebsiteLogFile -SiteName $script:websiteName @nonDefaultArgs -WhatIf
+        Set-CIisAppPoolProcessModel -AppPoolName $script:appPoolName @nonDefaultArgs -WhatIf
         ThenHasDefaultValues
     }
 
     It 'should support WhatIf when resetting all values back to defaults' {
-        Set-CIisWebsiteLogFile -SiteName $script:websiteName @nonDefaultArgs
+        Set-CIisAppPoolProcessModel -AppPoolName $script:appPoolName @nonDefaultArgs
         ThenHasValues $nonDefaultArgs
-        Set-CIisWebsiteLogFile -SiteName $script:websiteName -WhatIf
+        Set-CIisAppPoolProcessModel -AppPoolName $script:appPoolName -WhatIf
         ThenHasValues $nonDefaultArgs
     }
 
     It 'should change values and reset to defaults' {
-        Set-CIisWebsiteLogFile -SiteName $script:websiteName @nonDefaultArgs -ErrorAction Ignore
+        Set-CIisAppPoolProcessModel -AppPoolName $script:appPoolName @nonDefaultArgs -ErrorAction Ignore
         ThenHasValues $nonDefaultArgs
 
         $someArgs = @{
-            Directory = "C:\logs";
-            Period = [LoggingRolloverPeriod]::Weekly;
+            'UserName' = 'fubarsnafu';
+            'Password' = (ConvertTo-SecureString -String 'ufansrabuf' -AsPlainText -Force);
         }
-        Set-CIisWebsiteLogFile -SiteName $script:websiteName @someArgs
+        Set-CIisAppPoolProcessModel -AppPoolName $script:appPoolName @someArgs
         ThenHasValues $someArgs
     }
 
     It 'should change default settings' {
-        Set-CIisWebsiteLogFile -AsDefaults @nonDefaultArgs
+        Set-CIisAppPoolProcessModel -AsDefaults @nonDefaultArgs
         ThenDefaultsSetTo @nonDefaultArgs
     }
 }